feat(licenses): Add GPL-3 license option
This commit adds the GPL-3 license option to the license selection dropdown.

The GPL-3 license is a free and open source software license that provides users with the freedom to use, modify, and distribute software. It is a popular license for software that is intended to be used by a wide range of users, including commercial users.

By adding the GPL-3 license option, we are making it easier for users to choose the license that best suits their needs.